requirements.txt
tensorflow==2.0.0b0
pandas
numpy
tqdm
scikit-learn==0.21.3
matplotlib
nltk==3.4.5
spacy==2.2.3
pyLDAvis==2.1.2
gensim==3.8.1
pytest
pytest-mpl
pytest-mock
PyYAML
click